+++ src/felupe/thermal/__init__.py
@@ -1,7 +1,9 @@
 from ._solidbody_thermal import SolidBodyThermal
+from ._solidbody_thermal_convection import SolidBodyThermalConvection
 from ._time_step import TimeStep
 
 __all__ = [
     "SolidBodyThermal",
+    "SolidBodyThermalConvection",
     "TimeStep",
 ]