+++ tests/test_thermal.py
@@ -25,6 +25,10 @@ def test_thermal():
     temperature = fem.Field(region, dim=1)
     field = fem.FieldContainer([temperature])
 
+    region_convection = fem.RegionQuadBoundary(mesh, mask=mesh.y == 1.0)
+    temperature_convection = fem.Field(region_convection, dim=1)
+    field_convection = fem.FieldContainer([temperature_convection])
+
     boundaries = fem.BoundaryDict(
         left=fem.Boundary(temperature, fx=0),
         right=fem.Boundary(temperature, fx=1),
@@ -47,16 +51,21 @@ def test_thermal():
         thermal_conductivity=1.0,  # W/(m*K)
     )
 
-    solid.assemble.vector()
-    solid.assemble.vector(field)
+    convection = fem.thermal.SolidBodyThermalConvection(
+        field=field_convection,
+        coefficient=1.0,  # W/(m^2*K)
+        temperature=10.0,  # K
+    )
 
-    solid.assemble.matrix()
+    solid.assemble.vector(field)
     solid.assemble.matrix(field)
+    convection.assemble.vector(field)
+    convection.assemble.matrix(field)
 
     time = fem.thermal.TimeStep([solid])
     table = fem.math.linsteps([0, 1], num=2)
-    ramp = {boundaries["right"]: 10 * table, time: 0.1 * table}
-    step = fem.Step(items=[time, solid], ramp=ramp, boundaries=boundaries)
+    ramp = {boundaries["right"]: 10 * table, time: 0.1 * table, convection: 100 * table}
+    step = fem.Step(items=[time, solid, convection], ramp=ramp, boundaries=boundaries)
     job = fem.Job(steps=[step]).evaluate(
         filename="result.xdmf",  # result file for Paraview
         point_data={"Temperature": lambda field, substep: temperature.values},